# Speed-up factoring powers of small primes