### Before

```python
In [1]: import galois

In [2]: %time galois.perfect_power(2**471)
CPU times: user 3.03 s, sys: 0 ns, total: 3.03 s
Wall time: 3.03 s
Out[5]: (2, 471)

In [3]: %time galois.perfect_power(2**571)
# Takes forever....
```

### After

```python
In [1]: import galois

In [2]: %time galois.perfect_power(2**471)
CPU times: user 7 µs, sys: 14 µs, total: 21 µs
Wall time: 23.8 µs
Out[2]: (2, 471)

In [3]: %time galois.perfect_power(2**571)
CPU times: user 25 µs, sys: 0 ns, total: 25 µs
Wall time: 26.5 µs
Out[3]: (2, 571)
```